Final try to get multi-threading working
No luck. Had more of a proper look into it and it seems like some people have got it to work, but does not seem viable to implement. After trying for ~4 days, I have been unable to get it to work. Will have to discuss with team at later date.